+++ src/game/definitions/actors/index.ts
@@ -1,3 +1,4 @@
 export * from './ActorDefinitions';
 export * from './DialogueDefinitions';
 export * from './WorldInteractionDefinitions';
+export * from './EngineAlphaDefinitions';